move failurereasonenum and related types to shared modules
