+++ packages/grz-common/src/grz_common/utils/types.py
