+++ packages/grz-pydantic-models/src/grz_pydantic_models/status/__init__.py
@@ -1,3 +1,3 @@
-from .v1 import SubmissionStatusEntry, SubmissionStatusReport
+from .v1 import FailureReasonEnum, SubmissionStatusEntry, SubmissionStatusReport
 
-__all__ = ["SubmissionStatusEntry", "SubmissionStatusReport"]
+__all__ = ["FailureReasonEnum", "SubmissionStatusEntry", "SubmissionStatusReport"]
